Survey Validation - Cont.
Survey now finds highest count and appends the image correlating to it.
It also shows the user his or her profile info again and a link to
download the image via a relative link.
Next is to show the user all the answers that were entered in from the
survey.