Updates for group.
Tried cleaning up the code just a bit and moved a few files around.

However, the functionality is still all there, new users can be
register the site and a table of the correlating name is created as
well (via code). It then takes the newly register user to the profile
page and checks to make sure a session has been created - if not it
returns them to the register.php page.

Also added to the profile page a dropdown of many of the cellphone
carrier providers so that the person can select one when adding a new
person to their database.