Added Functionality For Adding Contact
Adding contacts via the profile page now works and sends the data the
the SQL DB storing it for later.
An alert is shown when it is added successfully (with the corresponding
phone number going into the DB).
Site also can login or signup users. It will also check to see if
someone is already logged in when accessing the profile page.