Backend PHP and MySQL DB Complete
Assignment now allows for a user to login, register, and logout.
Upon registering it adds to the user table a username; it also creates
a new table with the corresponding username.
All pages will verify that the user is logged in via a session on
either register or login.
Profile Page displays (currently) a table with the session values of
the user logged in.
The create page will allow the user to send a message to all contacts
associated with that table username.
The add page will allow the user to add to it a new person to the
database (to later be texted)
The contact list pulls out all the contacts in the corresponding table
and presents it on the page in HTML - via the get_contacts.php file
(this file pulls all the data into a single array that contains an
array at each position of the index).
Styling and a few other items can be added, but all the functionality
for the “backend” and database work is complete.